#291 Refactor Call Catalog into single-responsibility modules
Pure structural refactor, no behavior change. Breaks the 567-line
CatalogCallList and the inline pod filter into focused modules:

- formatTime.js: time-formatting helpers
- StaticWaveform.jsx / SpectrogramPanel.jsx: presentational pieces
- CallRow.jsx: the call row (collapsed row + expanded panel)
- useCatalogPlayer.js: playback state + handlers (custom hook)
- PodFilter.jsx: the pod filter tab bar
- CallCatalogSection.jsx: owns activePod; title/description/filter/list
- catalog.jsx slimmed to a thin page

Also removes the dead `export { CALLS }` re-export from CatalogCallList
(no importer; catalog.jsx imports CALLS from callsData directly).
