+++ src/components/Catalog/SpectrogramPanel.jsx
@@ -1,0 +1,24 @@
+import { Box } from '@mui/material'
+
+export default function SpectrogramPanel({ src, alt, maxWidth }) {
+  return (
+    <Box
+      aria-hidden={src ? undefined : 'true'}
+      sx={{
+        width: { xs: '100%', sm: maxWidth },
+        maxWidth,
+        overflow: 'hidden',
+        visibility: src ? 'visible' : 'hidden',
+      }}
+    >
+      {src && (
+        <Box
+          component="img"
+          src={src}
+          alt={alt}
+          sx={{ width: '100%', height: 'auto', display: 'block' }}
+        />
+      )}
+    </Box>
+  )
+}
